+++ fontend/src/components/AddLink/AddLink.scss
@@ -81,3 +81,8 @@
 }
 
 
+@media only screen and (max-width: 768px){
+    .AddLink{
+        width: 320px;
+    }
+}
