+++ fontend/src/components/Background/Background.scss
@@ -4,7 +4,7 @@
 .bgTheme {
     background-color: $white;
     border-radius: 25px;
-    width: 60%;
+    width: 70%;
     // display: flex;
     .bgTheme-items {
         display: grid;
