+++ fontend/src/components/ButtonLink/ButtonLink.scss
@@ -3,5 +3,5 @@
 .buttonLink{
     background-color: $white;
     border-radius: 25px;
-    width: 60%;
+    width: 70%;
 }
