+++ fontend/src/components/LinksItem/LinksItem.scss
@@ -156,3 +156,8 @@
         opacity: 0;
     }
 }
+@media only screen and (max-width: 768px){
+    .LinksItem{
+        width: 320px;
+    }
+}
