+++ fontend/src/components/Preview/PreView.scss
@@ -2,6 +2,7 @@
 @import '../../base/variable';
 
 .PreView-wapper {
+    // background-color: aqua;
     width: 100%;
     display: flex;
     justify-content: center;
