+++ fontend/src/components/TempProfile/TempProfile.scss
@@ -5,7 +5,7 @@
     flex-direction: column;
     background-color: $white;
     border-radius: 25px;
-    width: 60%;
+    width: 70%;
     margin: 0 auto;
     .tempProfile-item {
         display: flex;
