+++ fontend/src/components/TempTheme/TempTheme.scss
@@ -4,7 +4,7 @@
 .tempTheme {
     background-color: $white;
     border-radius: 25px;
-    width: 60%;
+    width: 70%;
     display: flex;
 
     .tempTheme-items {
