+++ fontend/src/pages/links/Links.jsx
@@ -21,20 +21,19 @@ function Links() {
         setIsAddLink(true);
     };
     const dispatch = useDispatch()
-    useEffect(()=>{
-        const fecthTheme = async() =>{
-            try {
-               const res = await http.get(`/card/v1/${currentUser._id}`) 
-               dispatch(themeSuccess(res.data))
-            } catch (error) {
-                console.log(error.message);
-            }
-        }
-        fecthTheme()
-    },[currentUser._id])
+    // useEffect(()=>{
+    //     const fecthTheme = async() =>{
+    //         try {
+    //            const res = await http.get(`/card/v1/${currentUser._id}`) 
+    //            dispatch(themeSuccess(res.data))
+    //         } catch (error) {
+    //             console.log(error.message);
+    //         }
+    //     }
+    //     fecthTheme()
+    // },[currentUser._id])
     return (
         <div className="Links">
-            <div className="Links-left">
                 <div className="Links-contents">
                     {!isAddLink && (
                         <button className="Links-left-btn" onClick={handleAddLink}>
@@ -50,11 +49,7 @@ function Links() {
                         <LinksItems onChange={onChange} />
                     </section>
                 </div>
-            </div>
-            {/* preview */}
-            <div className="Links-right">
-                <PreView  />
-            </div>
+            
         </div>
     );
 }
