+++ fontend/src/pages/tempate/Template.scss
@@ -1,18 +1,16 @@
 @import '../../base/color';
 @import '../../base/variable';
 .template {
-    display: flex;
     background-color: $bg_template;
-    height: 100%;
+    // height: 100%;
     padding-bottom: 50px;
-    .template-left {
-        padding-top: 180px;
-        flex: 3;
+    padding-top: 180px;
+    // width: 100%;
         .template-item {
+            // width: 80$;
             display: flex;
             flex-direction: column;
             align-items: center;
-
             .tempProfile_title {
                 padding-top: 10px;
                 margin-bottom: 20px;
@@ -20,106 +18,6 @@
                 font-weight: 800;
             }
         }
-    }
-    .template-right {
-        // border-left: 0.5px solid rgba($color: $black, $alpha: 0.3);
-        display: flex;
-        position: sticky;
-        justify-content: center;
-        top: 0;
-        height: 10px;
-        padding-top: 180px;
-        width: 30%;
-    }
-    .review-btn {
-        display: none;
-    }
-}
-//mobile
-@media only screen and (max-width: 768px) {
-    .template {
-        display: block;
-        position: relative;
-
-        .template-left {
-            .template-item {
-                width: 100%;
-            }
-        }
-
-        .template-right {
-            padding-top: 130px;
-            width: 100%;
-            // background-color: aqua;
-            z-index: 9999;
-            display: none;
-            position: fixed;
-            // position: absolute;
-            width: 100%;
-            left: 50%;
-            top: 40%;
-            transform: translate(-50%, -50%);
-            background-color: rgba($color: $black, $alpha: 0.4);
-            height: 100%;
-            margin: 0 auto;
-            animation: fadeIn 300ms cubic-bezier(0.075, 0.82, 0.165, 1);
-        }
-        .review-btn {
-            z-index: 99999;
-            display: block;
-            position: fixed;
-            bottom: -1.5%;
-            left: 50%;
-
-            transform: translate(-50%, -50%);
-            // top: 10%;
-            @include btn;
-            // padding: 10px;
-            background-color: $black;
-            color: $white;
-            padding: 12px;
-            font-size: 12px;
-            font-weight: 600;
-            border-radius: 10px;
-            // animation: tilt-shaking 100ms;
-            // animation: 3s ease-in 4s infinite reverse both running tilt-shaking;
-        }
-        .close {
-            background-color: var(--gray-6);
-            margin-bottom: 15px;
-            border-radius: 50%;
-            animation: none;
-        }
-    }
-}
-//table
-@media only screen and (min-width: 769px) and (max-width: 1024px) {
-}
-@keyframes fadeIn {
-    from {
-        opacity: 0;
-        top: 0%;
-    }
-    to {
-        opacity: 1;
-        top: 40%;
-    }
+  
 }
 
-@keyframes tilt-shaking {
-    0% {
-        transform: rotate(0deg);
-    }
-    25% {
-        transform: rotate(2deg);
-    }
-    50% {
-        transform: rotate(0eg);
-    }
-    75% {
-        transform: rotate(-2deg);
-    }
-    100% {
-        transform: rotate(0deg);
-    }
-}
